// src/core/enums/role.enum.ts
export enum Role {
  SUPER = 'SUPER', // Super Admin
  PRINCIPAL = 'PRINCIPAL', //Jefes del negocio
  ADMIN = 'ADMIN', //Jefe de Oficinas
  MANAGER = 'MANAGER', //Jefe de departamentos
  SUPERVISOR = 'SUPERVISOR', //Jefe de equipo
  AGENT = 'AGENT', // Team Members
  USER = 'USER', // Online Client
}